Upgrades various packages, fixes interface imports, switches completely to pages folder instead of app.vue.
