+++ .gitignore
@@ -97,4 +97,3 @@ config/
 !config/i18n.config.ts
 locales/
 !locales/en.json.example
-pages/
