+++ composables/useContent.ts
@@ -1,4 +1,4 @@
-import { MediaArguments } from '~/interfaces/Content';
+import type { MediaArguments } from '~/interfaces/Content';
 
 export const useMediaHandler = () => {
   function toImageUrl(url: string, {
