+++ composables/useMenu.ts
@@ -1,4 +1,4 @@
-import { HeaderMenu, MenuDepartment } from '~/interfaces/menu';
+import type { HeaderMenu, MenuDepartment } from '~/interfaces/menu';
 
 export const useMenuContent = () => useState<HeaderMenu>('menuContent', () => ({} as HeaderMenu));
 export const useSelectedMenuDepartment = () => useState<MenuDepartment | undefined>('selectedMenuDepartment', () => (undefined));
