--- layouts/blog-category.vue
@@ -1,5 +1,0 @@
-<template>
-  <LayoutsDefaultLayout>
-    <slot />
-  </LayoutsDefaultLayout>
-</template>
