+++ nuxt.config.ts
@@ -6,7 +6,7 @@ export default defineNuxtConfig({
   telemetry: false,
 
   // Disable servenr-side rendering: https://go.nuxtjs.dev/ssr-mode
-  // ssr: false, // nuxtjs/i18n is not working with ssr yet.
+  ssr: false, // nuxtjs/i18n is not working with ssr yet.
 
   runtimeConfig: {
     public: {
@@ -75,15 +75,6 @@ export default defineNuxtConfig({
     accessToken: process.env.STORYBLOK_API_TOKEN,
     useApiClient: true,
     apiOptions: { cache: { type: 'memory' } },
-  },
-
-  postcss: {
-    plugins: {
-      'postcss-import': {},
-      autoprefixer: {
-        overrideBrowserslist: ['last 2 versions', 'Firefox ESR', '> 1%', 'ie >= 8', 'iOS >= 8', 'Android >= 4'],
-      },
-    },
   },
 
   vite: {
