+++ package.json
@@ -11,21 +11,22 @@
     "storyblok-logout": "storyblok logout"
   },
   "devDependencies": {
-    "@nuxtjs/eslint-config-typescript": "^10.0.0",
-    "@typescript-eslint/eslint-plugin": "^5.26.0",
-    "@typescript-eslint/parser": "^5.26.0",
-    "eslint": "^8.16.0",
-    "eslint-plugin-vue": "^9.0.1",
-    "nuxt": "^3.0.0",
-    "storyblok": "^3.16.1"
+    "@nuxt/devtools": "1.0.8",
+    "@nuxtjs/eslint-config-typescript": "12.1.0",
+    "@typescript-eslint/eslint-plugin": "6.20.0",
+    "@typescript-eslint/parser": "6.20.0",
+    "eslint": "8.56.0",
+    "eslint-plugin-vue": "9.20.1",
+    "nuxt": "3.9.3",
+    "storyblok": "3.26.0"
   },
   "dependencies": {
-    "@nuxtjs/i18n": "npm:@nuxtjs/i18n-edge",
-    "@storyblok/nuxt": "^5.0.0",
-    "eslint-config-airbnb-base": "^15.0.0",
-    "nuxt-jsonld": "^2.0.6",
-    "sass": "^1.54.9",
-    "typescript": "^4.7.2",
-    "vite-plugin-eslint": "^1.6.1"
+    "@nuxtjs/i18n": "8.0.0",
+    "@storyblok/nuxt": "6.0.4",
+    "eslint-config-airbnb-base": "15.0.0",
+    "nuxt-jsonld": "2.0.8",
+    "sass": "1.70.0",
+    "typescript": "5.3.3",
+    "vite-plugin-eslint": "1.8.1"
   }
 }
