+++ storyblok/SbContentPage.vue
@@ -5,6 +5,17 @@
       required: true,
     },
   });
+  const pageMeta = usePageMeta();
+
+  const pageClasses = computed(() => {
+    const classes = [];
+
+    if (pageMeta.value.seamless_header) {
+      classes.push('content-page--seamless-header');
+    }
+
+    return classes;
+  });
 </script>
 
 <template>
@@ -12,6 +23,7 @@
     v-if="blok.content && Array.isArray(blok.content) && blok.content.length > 0"
     v-editable="blok"
     class="sb-content-page"
+    :class="pageClasses"
   >
     <component
       :is="$resolveStoryBlokComponent(block)"
